improve emon_fina module, move FinaReader and MetaData classe's to new fina_reader module
